Linking books to user (#26)
* wip

cc

* adding the logic to logout

* adding the authentication middleware

* adding the userId when adding a book

* fetching only books from the user

* clearing the book-user relationship when deleting

* wip

* adding the auto logout

+++ api/controllers/controllers-myShelf.js
@@ -1,7 +1,10 @@
 const Book = require('../models/book');
+const User = require('../models/user');
 
 exports.getBooks = (req, res, next) => {
-  Book.find()
+  const userId = req.userId;
+
+  Book.find({ userId: userId })
     .then((books) => {
       res.status(200).json({
         message: 'Fetched books successfully.',
@@ -20,12 +23,19 @@ exports.addNewBook = (req, res, next) => {
     title,
     author,
     image,
+    userId: req.userId,
   });
 
   newBook
     .save()
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.books.push(newBook);
+      return user.save();
+    })
+    .then(() => {
       res.status(200).json({ message: 'New book added' });
     })
     .catch((error) => console.log(error));
@@ -34,26 +44,27 @@ exports.addNewBook = (req, res, next) => {
 exports.deleteBook = async (req, res, next) => {
   const { bookId } = req.params;
 
-  // try {
-  //   const book = await Book.findById(bookId);
-
-  //   if (!book) throw new Error('No book found.');
-
-  //   await Book.findByIdAndRemove(bookId);
-
-  //   res.status(200).json({ message: 'Book deleted' });
-  // } catch (error) {
-  //   console.error(error);
-  // }
-
   Book.findById(bookId)
     .then((book) => {
       if (!book) {
-        throw new Error('No book found.');
+        const error = new Error('No book found.');
+        error.statusCode = 404;
+        throw error;
       }
-      //checked if the user is logged in
 
+      if (book.userId.toString() !== req.userId) {
+        const error = new Error('Not Authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
       return Book.findByIdAndRemove(bookId);
+    })
+    .then(() => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.books.pull(bookId);
+      return user.save();
     })
     .then(() => {
       res.status(200).json({ message: 'Book deleted.' });
